Standardize syntax, update games accordingly

+++ src/ludax/environment.py
@@ -57,7 +57,7 @@ class LudaxEnvironment():
             current_player=temp_current_player,
             phase_idx=jnp.int8(0),
             phase_step_count=jnp.int8(0),
-            previous_actions=jnp.int8([-1, -1, -1]),
+            previous_actions=jnp.int16([-1, -1, -1]),
         )
 
         # Initialize the board using the game rules
